New Indicator
Changed from old indicator to one reminiscent of kakitani's old website
