+++ package.json
@@ -31,7 +31,6 @@
     "react-native-element-dropdown": "^2.10.2",
     "react-native-gesture-handler": "~2.16.1",
     "react-native-get-random-values": "^1.11.0",
-    "react-native-indicators": "^0.17.0",
     "react-native-linear-gradient": "^2.8.3",
     "react-native-pager-view": "6.3.0",
     "react-native-reanimated": "~3.10.1",
